LJPS-53 added validation at service level

+++ backend/service/src/LearningJourneyService.js
@@ -13,6 +13,9 @@ function learningJourneyService(database) {
   }
 
   learningJourneyService.deleteLearningJourney = async (learningJourneyId) => {
+    if(!learningJourneyId){
+      return false 
+    }
     try {
       return await database.deleteLearningJourney(learningJourneyId)
     }
